fix: revise cvector -> _vector to fit the validate.sh requirement
